requirements.txt
annotated-types==0.7.0
anyio==4.13.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.11.1
attrs==26.1.0
better-profanity==0.7.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
cloudinary==1.44.1
cryptography==46.0.5
Deprecated==1.3.1
dj-rest-auth==7.2.0
Django==6.0.3
django-allauth==65.15.0
django-anymail==14.0
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.3.1
googleapis-common-protos==1.74.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
joblib==1.5.3
Markdown==3.10.2
numpy==2.4.4
oauthlib==3.3.1
pillow==12.2.0
proto-plus==1.27.2
protobuf==7.34.1
psycopg2-binary==2.9.11
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.12.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
requests==2.32.5
requests-oauthlib==2.0.0
scikit-learn==1.8.0
scipy==1.17.1
setuptools==82.0.1
six==1.17.0
sqlparse==0.5.5
standardwebhooks==1.0.1
svix==1.90.0
threadpoolctl==3.6.0
types-Deprecated==1.3.1.20260408
types-python-dateutil==2.9.0.20260408
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
wrapt==2.1.2
